Added the required memory property for AttachedElementCells, insofar as needed for Skeleton View validation.
Next is figuring out how the submit button will work going forward and optimizing the Editor view workflow with respect to possible early completion.

Possible early completion: after successfully completing the Skeleton View with a formula of methane (CH4), the student is done. Ideally, in such a case, the Octets segment wouldn't even need to be added to the main segmented control.