Implemented the Bonds View transition. Next is finishing the validation.
The focus of the Bonds View is that, by the nature of reaching this View, the student's structure likely has too many lone pairs and not enough bonds. Removing lone pairs and adding bonds appropriately to correct the difference in suggested vs. actual valence totals is what this View tries to accomplish.